Move convert scalars into their own file in recordset engine.
