+++ rust/experimental/query_engine/engine-recordset/src/scalars/mod.rs
@@ -1,8 +1,10 @@
+pub(crate) mod convert_scalar_expressions;
 pub(crate) mod math_scalar_expressions;
 pub(crate) mod parse_scalar_expressions;
 pub(crate) mod scalar_expressions;
 pub(crate) mod temporal_scalar_expressions;
 
+pub use convert_scalar_expressions::*;
 pub use math_scalar_expressions::*;
 pub use parse_scalar_expressions::*;
 pub use scalar_expressions::*;
